Remove unused argument to Heap and make size size_t

+++ src/memory/MarkSweepHeap.h
@@ -7,7 +7,7 @@
 class MarkSweepHeap : public Heap<MarkSweepHeap> {
     friend class MarkSweepCollector;
 public:
-    MarkSweepHeap(long objectSpaceSize = 1048576);
+    MarkSweepHeap(size_t objectSpaceSize = 1048576);
     AbstractVMObject* AllocateObject(size_t size);
 private:
     vector<AbstractVMObject*>* allocatedObjects;
